prepmail update

+++ set 8/prepmail.sh
@@ -11,23 +11,31 @@ if [ "$#" -ne 2 ]; then
 fi
 
 # dir
-mkdir mail
+if [ -d "mail" ]; then
+    echo "'/mail' exists, continuing"
+else
+    mkdir mail
+fi
 
 # loop over exercises
 for i in `seq $1 $2`;
 do
-    # copy 
-    cp -r $i mail/
-    cd mail/$i 
-    # clean 
-    make clean
-    rm Makefile makeclass GIT_INFO
-    # prepare order.txt
-    order=$(tree -fi)
-    echo "$order" >> order.txt
-    sed -i 's|./||' order.txt
-    sed -i '1d' order.txt
-    sed -i '$ d' order.txt
-    sed -i '$ d' order.txt
-    cd ../..
+    if [ -d "mail/$i" ]; then
+        echo "folder $i exists! skipping."
+    else 
+        # copy 
+        cp -r $i mail/
+        cd mail/$i 
+        # clean 
+        make clean
+        rm Makefile makeclass GIT_INFO
+        # prepare order.txt
+        order=$(tree -fi)
+        echo "$order" >> order.txt
+        sed -i 's|./||' order.txt
+        sed -i '1d' order.txt
+        sed -i '$ d' order.txt
+        sed -i '$ d' order.txt
+        cd ../..
+    fi
 done
